Refactor: 게시글 목록을 MyPostsList라는 별도의 컴포넌트로 분리
